feat: [M3-7036] component cleanup
